[refract] : moving constant files to constant directory
