+++ lib/view/screens/Authentication/login_screen.dart
@@ -12,8 +12,8 @@ import 'package:wave/utils/constants/custom_fonts.dart';
 import 'package:wave/utils/constants/cutom_logo.dart';
 import 'package:wave/utils/device_size.dart';
 import 'package:wave/utils/enums.dart';
-import 'package:wave/utils/keys.dart';
-import 'package:wave/utils/preferences.dart';
+import 'package:wave/utils/constants/keys.dart';
+import 'package:wave/utils/constants/preferences.dart';
 import 'package:wave/utils/routing.dart';
 import 'package:wave/view/reusable_components/auth_textfield.dart';
