+++ lib/view/screens/Authentication/register_screen.dart
@@ -11,7 +11,7 @@ import 'package:wave/utils/constants/custom_colors.dart';
 import 'package:wave/utils/constants/custom_fonts.dart';
 import 'package:wave/utils/device_size.dart';
 import 'package:wave/utils/enums.dart';
-import 'package:wave/utils/keys.dart';
+import 'package:wave/utils/constants/keys.dart';
 import 'package:wave/utils/util_functions.dart';
 import 'package:wave/view/reusable_components/auth_textfield.dart';
